Update README with token rights
The current token rights aren't sufficient to get a travis token for a private repo. I tested what is minimally needed and came to the conclusion that without repo it doesn't work. Github removes the  "repo_deployment" and "repo:status" rights then, I guess because they overlap. I got the wallboard fully working with the token.